remove note

+++ mnmp-protocol/src/main/java/com/tarena/mnmp/enums/SendType.java
@@ -42,7 +42,7 @@ public enum SendType {
     }
 
     public static boolean notImmediately(int type) {
-        if (DELAY.type.intValue() == type || CYCLE.type.intValue() == type) {
+        if (DELAY.type == type || CYCLE.type == type) {
             return true;
         }
         return false;
